+++ lib/elasticity/emr.rb
@@ -4,8 +4,8 @@ module Elasticity
 
     attr_reader :aws_request
 
-    def initialize(aws_access_key_id=nil, aws_secret_access_key=nil, options = {})
-      @aws_request = Elasticity::AwsSession.new(aws_access_key_id, aws_secret_access_key, options)
+    def initialize(options = {})
+      @aws_request = Elasticity::AwsSession.new(options)
     end
 
     # Adds a new group of instances to the specified jobflow.  Elasticity maps a